Completed Value ADDED Features and Refactors

+++ HotelListingAPI/Controllers/CountriesV2Controller .cs
@@ -12,22 +12,22 @@ using AutoMapper;
 using HotelListingAPI.Contracts;
 using Microsoft.AspNetCore.Authorization;
 using HotelListingAPI.Exceptions;
-using HotelListingAPI.Models;
+using Microsoft.AspNetCore.OData.Query;
 
 namespace HotelListingAPI.Controllers
 {
     [Route("api/v{version:apiVersion}countries")]
     [ApiController]
-    [ApiVersion("1.0",Deprecated= true)]
-   
-    public class CountriesController : ControllerBase
+    [ApiVersion("2.0")]
+
+    public class CountriesV2Controller : ControllerBase
     {
         //private readonly HotelListingDbContext _context;
         private readonly IMapper _mapper;
         private readonly ICountriesRepository _countriesRepository;
         private readonly ILogger<CountriesController> _logger;
 
-        public CountriesController(IMapper mapper,ICountriesRepository countriesRepository,ILogger<CountriesController>logger)
+        public CountriesV2Controller(IMapper mapper,ICountriesRepository countriesRepository,ILogger<CountriesController>logger)
         {
            // _context = context;
             this._mapper = mapper;
@@ -35,8 +35,9 @@ namespace HotelListingAPI.Controllers
             this._logger = logger;
         }
 
-        // GET: api/Countries/getall
-        [HttpGet("GetAll")]
+        // GET: api/Countries
+        [HttpGet]
+        [EnableQuery]
         public async Task<ActionResult<IEnumerable<GetCountryDto>>> GetCountries()
         {
           //if (_context.Countries == null)
@@ -45,20 +46,10 @@ namespace HotelListingAPI.Controllers
           //}
             //select * from countries;
             // var countries=await _context.Countries.ToListAsync();
-            //var countries = await _countriesRepository.GetAllAsync();
-            //var records=_mapper.Map<List<GetCountryDto>>(countries);
-            var countries= await _countriesRepository.GetAllAsync<GetCountryDto>();
-            return Ok(countries);
+            var countries = await _countriesRepository.GetAllAsync();
+            var records=_mapper.Map<List<GetCountryDto>>(countries);
+            return Ok(records);
             //return await _context.Countries.ToListAsync();
-        }
-        // GET: api/Countries/?StartIndex=0&pagesize=2&pagenumber=1
-        [HttpGet]
-        public async Task<ActionResult<PagedResult<IEnumerable<GetCountryDto>>>> GetPagedCountries([FromQuery]QueryParameters queryParameters)
-        {
-           
-            var pagedCountriesResult = await _countriesRepository.GetAllAsync<GetCountryDto>(queryParameters);  
-            return Ok(pagedCountriesResult);
-            
         }
 
         // GET: api/Countries/5
@@ -71,19 +62,17 @@ namespace HotelListingAPI.Controllers
             //}
             //var country = await _context.Countries.Include(q =>q.Hotels)
             //    .FirstOrDefaultAsync(q => q.Id == id);
-            //var country = await _countriesRepository.GetDetails(id);
-
-
-            //if (country == null)
-            //{
-            //    //_logger.LogWarning($"record not found in {nameof(GetCountry)} with id: {id} ");
-            //    //return NotFound();
-            //    throw new NotFoundException(nameof(GetCountry),id);
-            //}
-            //var countryDto = _mapper.Map<CountryDto>(country);
-            //return Ok(countryDto);
             var country = await _countriesRepository.GetDetails(id);
-                return Ok(country);
+            
+
+            if (country == null)
+            {
+                //_logger.LogWarning($"record not found in {nameof(GetCountry)} with id: {id} ");
+                //return NotFound();
+                throw new NotFoundException(nameof(GetCountry),id);
+            }
+            var countryDto = _mapper.Map<CountryDto>(country);
+            return Ok(countryDto);
         }
 
         // PUT: api/Countries/5
@@ -99,18 +88,18 @@ namespace HotelListingAPI.Controllers
 
             // _context.Entry(country).State = EntityState.Modified;
             //var country = await _context.Countries.FindAsync(id);
-            //var country = await _countriesRepository.GetAsync(id);
+            var country = await _countriesRepository.GetAsync(id);
 
-            //if (country == null)
-            //{
-            //    //return NotFound();
-            //    throw new NotFoundException(nameof(GetCountry), id);
-            //}
-            //_mapper.Map(updateCountryDto, country);
+            if (country == null)
+            {
+                //return NotFound();
+                throw new NotFoundException(nameof(GetCountry), id);
+            }
+            _mapper.Map(updateCountryDto, country);
 
             try
             {
-                await _countriesRepository.UpdateAsync(id,updateCountryDto);
+                await _countriesRepository.UpdateAsync(country);
             }
             catch (DbUpdateConcurrencyException)
             {
@@ -131,26 +120,22 @@ namespace HotelListingAPI.Controllers
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
         [Authorize]
-        public async Task<ActionResult<CountryDto>> PostCountry(CreateCountryDto createCountryDto)
+        public async Task<ActionResult<Country>> PostCountry(CreateCountryDto createCountryDto)
         {
             //var country = new Country()
             //{
             //    Name = createCountry.Name,
             //    ShortName = createCountry.ShortName
             //};
-            //var country=_mapper.Map<Country>(createCountryDto);
+            var country=_mapper.Map<Country>(createCountryDto);
 
-            //if (_countriesRepository.GetAsync== null)
-            //{
-            //    return Problem("Entity set 'HotelListingDbContext.Countries'  is null.");
-            //}
-            //  //_context.Countries.Add(country);
-            //  //await _context.SaveChangesAsync();
-            // await  _countriesRepository.AddAsync(country);
-            // // await _countriesRepository.UpdateAsync(country);
-            var country = await _countriesRepository.AddAsync<CreateCountryDto, GetCountryDto>(createCountryDto);
+         
+            //_context.Countries.Add(country);
+            //await _context.SaveChangesAsync();
+           await  _countriesRepository.AddAsync(country);
+           // await _countriesRepository.UpdateAsync(country);
 
-            return CreatedAtAction(nameof(GetCountry), new { id = country.Id }, country);
+            return CreatedAtAction("GetCountry", new { id = country.Id }, country);
         }
 
         // DELETE: api/Countries/5
@@ -162,12 +147,12 @@ namespace HotelListingAPI.Controllers
             //{
             //    return NotFound();
             //}
-            //var country = await _countriesRepository.GetAsync(id);
-            //if (country == null)
-            //{
-            //    //return NotFound();
-            //    throw new NotFoundException(nameof(GetCountry), id);
-            //}
+            var country = await _countriesRepository.GetAsync(id);
+            if (country == null)
+            {
+                //return NotFound();
+                throw new NotFoundException(nameof(GetCountry), id);
+            }
 
            await  _countriesRepository.DeleteAsync(id);
             //await _countriesRepository.UpdateAsync(country);
